+++ src/main/java/btw/community/classicaddon/ClassicAddon.java
@@ -63,6 +63,10 @@ public class ClassicAddon extends BTWAddon {
 	public static boolean guaranteedSeedDrop;
 	public static boolean intentionalHungerRegenOffset;
 	public static boolean canBabyAnimalEatLooseFood;
+	public static boolean hempSeedDropFromTallGrass;
+	public static boolean hardcoreSpawnToggle;
+	public static boolean hcHoofsiesToggle;
+	public static boolean strongerHoofsiesToggle;
 	public static boolean passableLeaves;
 	public static boolean vanillaifyBuckets;
 	public static boolean yeetTooExpensive;
@@ -108,6 +112,9 @@ public class ClassicAddon extends BTWAddon {
 		guaranteedSeedDrop = Boolean.parseBoolean(propertyValues.get("GuaranteedSeedDrop"));
 		intentionalHungerRegenOffset = Boolean.parseBoolean(propertyValues.get("IntentionalHungerRegenOffset"));
 		canBabyAnimalEatLooseFood = Boolean.parseBoolean(propertyValues.get("CanBabyAnimalEatLooseFood"));
+		hempSeedDropFromTallGrass = Boolean.parseBoolean(propertyValues.get("HempSeedDropFromTallGrass"));
+		hcHoofsiesToggle = Boolean.parseBoolean(propertyValues.get("HCHoofsiesToggle"));
+		strongerHoofsiesToggle = Boolean.parseBoolean(propertyValues.get("StrongerHoofsies"));
 		synchronizedConfigProperties.forEach(((propertyName, configProperty) -> {
 			configProperty.setInternalValue(propertyValues.get(propertyName));
 			configProperty.resetExternalValue();
@@ -124,11 +131,17 @@ public class ClassicAddon extends BTWAddon {
 		this.registerProperty("AnimageddonToggle", "False", "A toggle for BTW Animageddon. Turning this off will disable animal hunger, makes sheep's wool insta-grow when grazing one grass, wolves need to be fed once to shit.");
 		this.registerProperty("ChickenJockeyToggle", "False", "This toggles spawning of buggy chicken jockeys.");
 		this.registerProperty("IntentionalHungerRegenOffset", "True", "This shifts the regen stop region to be below 8.6 shanks instead of below 9 shanks.\n# This makes regen feel much more consistent, even if internally, it may not exactly match up.");
-		this.registerProperty("GuaranteedSeedDrop", "True", "This makes sure that seeds will always drop, no matter the growth stage, just like in modern vanilla.");
+		this.registerProperty("GuaranteedSeedDrop", "True", "This makes sure that crop seeds will always drop, no matter the growth stage, just like in modern vanilla.");
 		this.registerProperty("CanBabyAnimalEatLooseFood", "False", "A toggle to re-introduce the bug with baby animal eating off of ground. This only works while Animageddon is turned off.");
-		this.registerSynchronizedProperty("PassableLeaves", "True",
+		this.registerProperty("HempSeedDropFromTallGrass", "True", "This toggles the 1% drop chance for hemp seeds from tall grass");
+		this.registerProperty("HCHoofsiesToggle", "False", "This toggles the HC Hoofsies mechanic from BTW. This only affects the True Classic difficulty.");
+		this.registerProperty("StrongerHoofsies", "False", "Toggling this on makes kicking animals deal 7 HP. This only affects the True Classic difficulty.");
+		this.registerSynchronizedProperty("PassableLeaves", "False",
 				string -> passableLeaves = Boolean.parseBoolean(string),
 				" *** SYNCHRONIZED PROPERTIES ***\n\n# This toggles the passable leaves functionality.");
+		this.registerSynchronizedProperty("HardcoreSpawnToggle", "False",
+				string -> hardcoreSpawnToggle = Boolean.parseBoolean(string),
+				"This toggles the HC Spawn mechanic from BTW. This only affects the True Classic difficulty.");
 		this.registerSynchronizedProperty("VanillaifyBuckets", "True",
 				string -> vanillaifyBuckets = Boolean.parseBoolean(string),
 				"This option re-introduces vanilla bucket mechanics. This makes screw pumps useless.");